requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
razorpay>=1.4
Pillow>=10.0

